backend/requirements-light.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI and NLP (lighter versions)
langchain==0.0.350
langchain-community==0.0.1
transformers==4.35.2
torch==2.1.1
numpy==1.24.3

# Web scraping
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0

# PDF processing
pdfplumber==0.10.3
Pillow==10.1.0

# Graph processing
networkx==3.2.1

# Data processing
pandas==2.1.4
python-dotenv==1.0.0

# CORS for frontend
python-cors==1.7.0 